Feat: started renaming entities
