+++ input_utxo.go
@@ -17,7 +17,7 @@ const (
 // UTXOInput references an unspent transaction output by the Transaction's ID and the corresponding index of the Output.
 type UTXOInput struct {
 	// The transaction ID of the referenced transaction.
-	TransactionID TransactionID `serix:"0,mapKey=transactionId"`
+	TransactionID SignedTransactionID `serix:"0,mapKey=transactionId"`
 	// The output index of the output on the referenced transaction.
 	TransactionOutputIndex uint16 `serix:"1,mapKey=transactionOutputIndex"`
 }
@@ -69,7 +69,7 @@ func (u *UTXOInput) Equals(other *UTXOInput) bool {
 }
 
 func (u *UTXOInput) Size() int {
-	// InputType + TransactionID + TransactionOutputIndex
+	// InputType + SignedTransactionID + TransactionOutputIndex
 	return serializer.SmallTypeDenotationByteSize + SlotIdentifierLength + OutputIndexLength
 }
