+++ transaction.go
@@ -1,240 +1,268 @@
 package iotago
 
 import (
-	"fmt"
+	"golang.org/x/crypto/blake2b"
 
 	"github.com/iotaledger/hive.go/ierrors"
 	"github.com/iotaledger/hive.go/serializer/v2"
 )
 
+// TransactionType defines the type of transaction.
+type TransactionType = byte
+
 const (
-	// TransactionIDLength defines the length of a Transaction ID.
-	TransactionIDLength = SlotIdentifierLength
+	// TransactionNormal denotes a standard transaction essence.
+	TransactionNormal TransactionType = 2
+
+	// MinContextInputsCount defines the minimum amount of context inputs within a Transaction.
+	MinContextInputsCount = 0
+	// MaxContextInputsCount defines the maximum amount of context inputs within a Transaction.
+	MaxContextInputsCount = 128
+	// MaxInputsCount defines the maximum amount of inputs within a Transaction.
+	MaxInputsCount = 128
+	// MinInputsCount defines the minimum amount of inputs within a Transaction.
+	MinInputsCount = 1
+	// MaxOutputsCount defines the maximum amount of outputs within a Transaction.
+	MaxOutputsCount = 128
+	// MinOutputsCount defines the minimum amount of inputs within a Transaction.
+	MinOutputsCount = 1
+	// MinAllotmentCount defines the minimum amount of allotments within a Transaction.
+	MinAllotmentCount = 0
+	// MaxAllotmentCount defines the maximum amount of allotments within a Transaction.
+	MaxAllotmentCount = 128
+
+	// InputsCommitmentLength defines the length of the inputs commitment hash.
+	InputsCommitmentLength = blake2b.Size256
 )
 
 var (
-	// ErrMissingUTXO gets returned if an UTXO is missing to commence a certain operation.
-	ErrMissingUTXO = ierrors.New("missing utxo")
-	// ErrInputOutputSumMismatch gets returned if a transaction does not spend the entirety of the inputs to the outputs.
-	ErrInputOutputSumMismatch = ierrors.New("inputs and outputs do not spend/deposit the same amount")
-	// ErrManaOverflow gets returned when there is an under- or overflow in Mana calculations.
-	ErrManaOverflow = ierrors.New("under- or overflow in Mana calculations")
-	// ErrUnknownSignatureType gets returned for unknown signature types.
-	ErrUnknownSignatureType = ierrors.New("unknown signature type")
-	// ErrSignatureAndAddrIncompatible gets returned if an address of an input has a companion signature unlock with the wrong signature type.
-	ErrSignatureAndAddrIncompatible = ierrors.New("address and signature type are not compatible")
-	// ErrInvalidInputUnlock gets returned when an input unlock is invalid.
-	ErrInvalidInputUnlock = ierrors.New("invalid input unlock")
-	// ErrSenderFeatureNotUnlocked gets returned when an output contains a SenderFeature with an ident which is not unlocked.
-	ErrSenderFeatureNotUnlocked = ierrors.New("sender feature is not unlocked")
-	// ErrIssuerFeatureNotUnlocked gets returned when an output contains a IssuerFeature with an ident which is not unlocked.
-	ErrIssuerFeatureNotUnlocked = ierrors.New("issuer feature is not unlocked")
-	// ErrReturnAmountNotFulFilled gets returned when a return amount in a transaction is not fulfilled by the output side.
-	ErrReturnAmountNotFulFilled = ierrors.New("return amount not fulfilled")
-	// ErrInputOutputManaMismatch gets returned if Mana is not balanced across inputs and outputs/allotments.
-	ErrInputOutputManaMismatch = ierrors.New("inputs and outputs do not contain the same amount of Mana")
-	// ErrInputCreationAfterTxCreation gets returned if an input has creation slot after the transaction creation slot.
-	ErrInputCreationAfterTxCreation = ierrors.New("input creation slot after tx creation slot")
-	// ErrUnknownTransactionEssenceType gets returned for unknown transaction essence types.
-	ErrUnknownTransactionEssenceType = ierrors.New("unknown transaction essence type")
+	// ErrInvalidInputsCommitment gets returned when the inputs commitment is invalid.
+	ErrInvalidInputsCommitment = ierrors.New("invalid inputs commitment")
+	// ErrTxEssenceNetworkIDInvalid gets returned when a network ID within a Transaction is invalid.
+	ErrTxEssenceNetworkIDInvalid = ierrors.New("invalid network ID")
+	// ErrInputUTXORefsNotUnique gets returned if multiple inputs reference the same UTXO.
+	ErrInputUTXORefsNotUnique = ierrors.New("inputs must each reference a unique UTXO")
+	// ErrInputBICNotUnique gets returned if multiple inputs reference the same BIC.
+	ErrInputBICNotUnique = ierrors.New("inputs must each reference a unique BIC")
+	// ErrInputRewardInvalid gets returned if multiple reward inputs reference the same input index
+	// or if they reference an index greater than max inputs count.
+	ErrInputRewardInvalid = ierrors.New("invalid reward input")
+	// ErrMultipleInputCommitments gets returned if multiple commitment inputs are provided.
+	ErrMultipleInputCommitments = ierrors.New("there are multiple commitment inputs")
+	// ErrAccountOutputNonEmptyState gets returned if an AccountOutput with zeroed AccountID contains state (counters non-zero etc.).
+	ErrAccountOutputNonEmptyState = ierrors.New("account output is not empty state")
+	// ErrAccountOutputCyclicAddress gets returned if an AccountOutput's AccountID results into the same address as the State/Governance controller.
+	ErrAccountOutputCyclicAddress = ierrors.New("account output's AccountID corresponds to state and/or governance controller")
+	// ErrNFTOutputCyclicAddress gets returned if an NFTOutput's NFTID results into the same address as the address field within the output.
+	ErrNFTOutputCyclicAddress = ierrors.New("NFT output's ID corresponds to address field")
+	// ErrDelegationValidatorAddressZeroed gets returned if a Delegation Output's Validator address is zeroed out.
+	ErrDelegationValidatorAddressZeroed = ierrors.New("delegation output's validator address is zeroed")
+	// ErrOutputsSumExceedsTotalSupply gets returned if the sum of the output deposits exceeds the total supply of tokens.
+	ErrOutputsSumExceedsTotalSupply = ierrors.New("accumulated output balance exceeds total supply")
+	// ErrOutputAmountMoreThanTotalSupply gets returned if an output base token amount is more than the total supply.
+	ErrOutputAmountMoreThanTotalSupply = ierrors.New("an output's base token amount cannot exceed the total supply")
+	// ErrStorageDepositLessThanMinReturnOutputStorageDeposit gets returned when the storage deposit condition's amount is less than the min storage deposit for the return output.
+	ErrStorageDepositLessThanMinReturnOutputStorageDeposit = ierrors.New("storage deposit return amount is less than the min storage deposit needed for the return output")
+	// ErrStorageDepositExceedsTargetOutputAmount gets returned when the storage deposit condition's amount exceeds the target output's base token amount.
+	ErrStorageDepositExceedsTargetOutputAmount = ierrors.New("storage deposit return amount exceeds target output's base token amount")
+	// ErrMaxNativeTokensCountExceeded gets returned if outputs or transactions exceed the MaxNativeTokensCount.
+	ErrMaxNativeTokensCountExceeded = ierrors.New("max native tokens count exceeded")
 )
 
-var (
-	EmptyTransactionID = TransactionID{}
-)
+// TransactionSelector implements SerializableSelectorFunc for transaction types.
+func TransactionSelector(txType uint32) (*Transaction, error) {
+	var seri *Transaction
+	switch byte(txType) {
+	case TransactionNormal:
+		seri = &Transaction{}
+	default:
+		return nil, ierrors.Wrapf(ErrUnknownTransactionEssenceType, "type byte %d", txType)
+	}
 
-type TransactionID = SlotIdentifier
-
-// TransactionEssenceID is the ID of a TransactionEssence.
-type TransactionEssenceID = Identifier
-
-// TransactionIDs are IDs of transactions.
-type TransactionIDs []TransactionID
-
-// TransactionIDFromData returns a new TransactionID for the given data by hashing it with blake2b and appending the creation slot index.
-func TransactionIDFromData(creationSlot SlotIndex, data []byte) TransactionID {
-	return SlotIdentifierRepresentingData(creationSlot, data)
+	return seri, nil
 }
 
-type TransactionContextInputs ContextInputs[Input]
+// InputsCommitment is a commitment to the inputs of a transaction.
+type InputsCommitment = [InputsCommitmentLength]byte
 
-// Transaction is a transaction with its inputs, outputs and unlocks.
+type (
+	txEssenceContextInput  interface{ Input }
+	txEssenceInput         interface{ Input }
+	TxEssenceOutput        interface{ Output }
+	TxEssencePayload       interface{ Payload }
+	TxEssenceContextInputs = ContextInputs[txEssenceContextInput]
+	TxEssenceInputs        = Inputs[txEssenceInput]
+	TxEssenceOutputs       = Outputs[TxEssenceOutput]
+	TxEssenceAllotments    = Allotments
+)
+
+// Transaction is the essence part of a Transaction.
 type Transaction struct {
-	// The transaction essence, respectively the transfer part of a Transaction.
-	Essence *TransactionEssence `serix:"0,mapKey=essence"`
-	// The unlocks defining the unlocking data for the inputs within the Essence.
-	Unlocks Unlocks `serix:"1,mapKey=unlocks"`
+	// TransactionEssence is the essence of a Transaction.
+	*TransactionEssence `serix:"0"`
+	// The outputs of this transaction.
+	Outputs TxEssenceOutputs `serix:"1,mapKey=outputs"`
 }
 
-func (t *Transaction) Clone() Payload {
+// TransactionEssence is the essence part of a Transaction.
+type TransactionEssence struct {
+	// The network ID for which this essence is valid for.
+	NetworkID NetworkID `serix:"0,mapKey=networkId"`
+	// The slot index in which the transaction was created by the client.
+	CreationSlot SlotIndex `serix:"1,mapKey=creationSlot"`
+	// The commitment references of this transaction.
+	ContextInputs TxEssenceContextInputs `serix:"2,mapKey=contextInputs"`
+	// The inputs of this transaction.
+	Inputs TxEssenceInputs `serix:"3,mapKey=inputs"`
+	// The commitment to the referenced inputs.
+	InputsCommitment InputsCommitment `serix:"4,mapKey=inputsCommitment"`
+	// The optional accounts map with corresponding allotment values.
+	Allotments TxEssenceAllotments `serix:"5,mapKey=allotments"`
+	// The optional embedded payload.
+	Payload TxEssencePayload `serix:"6,optional,mapKey=payload"`
+}
+
+// Clone creates a copy of the Transaction.
+func (u *Transaction) Clone() *Transaction {
+	var payload TxEssencePayload
+	if u.Payload != nil {
+		payload = u.Payload.Clone()
+	}
+
 	return &Transaction{
-		Essence: t.Essence.Clone(),
-		Unlocks: t.Unlocks.Clone(),
+		TransactionEssence: &TransactionEssence{
+			NetworkID:        u.NetworkID,
+			CreationSlot:     u.CreationSlot,
+			ContextInputs:    u.ContextInputs.Clone(),
+			Inputs:           u.Inputs.Clone(),
+			InputsCommitment: u.InputsCommitment,
+			Allotments:       u.Allotments.Clone(),
+			Payload:          payload,
+		},
+		Outputs: u.Outputs.Clone(),
 	}
 }
 
-func (t *Transaction) PayloadType() PayloadType {
-	return PayloadTransaction
-}
-
-// OutputsSet returns an OutputSet from the Transaction's outputs, mapped by their OutputID.
-func (t *Transaction) OutputsSet(api API) (OutputSet, error) {
-	txID, err := t.ID(api)
+// SigningMessage returns the to be signed message.
+func (u *Transaction) SigningMessage(api API) ([]byte, error) {
+	essenceBytes, err := api.Encode(u)
 	if err != nil {
 		return nil, err
 	}
-	set := make(OutputSet)
-	for index, output := range t.Essence.Outputs {
-		set[OutputIDFromTransactionIDAndIndex(txID, uint16(index))] = output
-	}
+	essenceBytesHash := blake2b.Sum256(essenceBytes)
 
-	return set, nil
+	return essenceBytesHash[:], nil
 }
 
-// ID computes the ID of the Transaction.
-func (t *Transaction) ID(api API) (TransactionID, error) {
-	data, err := api.Encode(t)
+// Sign produces signatures signing the Transaction for every given AddressKeys.
+// The produced signatures are in the same order as the AddressKeys.
+func (u *Transaction) Sign(api API, inputsCommitment []byte, addrKeys ...AddressKeys) ([]Signature, error) {
+	if inputsCommitment == nil || len(inputsCommitment) != InputsCommitmentLength {
+		return nil, ErrInvalidInputsCommitment
+	}
+
+	copy(u.InputsCommitment[:], inputsCommitment)
+
+	signMsg, err := u.SigningMessage(api)
 	if err != nil {
-		return TransactionID{}, ierrors.Errorf("can't compute transaction ID: %w", err)
+		return nil, err
 	}
 
-	return TransactionIDFromData(t.Essence.CreationSlot, data), nil
-}
-
-// EssenceID computes the essence ID of the Transaction.
-func (t *Transaction) EssenceID(api API) (TransactionEssenceID, error) {
-	data, err := api.Encode(t.Essence)
-	if err != nil {
-		return TransactionEssenceID{}, ierrors.Errorf("can't compute transaction essence ID: %w", err)
-	}
-
-	return IdentifierFromData(data), nil
-}
-
-func (t *Transaction) Inputs() ([]*UTXOInput, error) {
-	references := make([]*UTXOInput, 0, len(t.Essence.Inputs))
-	for _, input := range t.Essence.Inputs {
-		switch castInput := input.(type) {
-		case *UTXOInput:
-			references = append(references, castInput)
-		default:
-			return nil, ErrUnknownInputType
+	sigs := make([]Signature, len(addrKeys))
+	signer := NewInMemoryAddressSigner(addrKeys...)
+	for i, v := range addrKeys {
+		sig, err := signer.Sign(v.Address, signMsg)
+		if err != nil {
+			return nil, err
 		}
+		sigs[i] = sig
 	}
 
-	return references, nil
+	return sigs, nil
 }
 
-func (t *Transaction) ContextInputs() (TransactionContextInputs, error) {
-	references := make(TransactionContextInputs, 0, len(t.Essence.ContextInputs))
-	for _, input := range t.Essence.ContextInputs {
-		switch castInput := input.(type) {
-		case *CommitmentInput, *BlockIssuanceCreditInput, *RewardInput:
-			references = append(references, castInput)
-		default:
-			return nil, ErrUnknownContextInputType
-		}
+func (u *Transaction) Size() int {
+	payloadSize := serializer.PayloadLengthByteSize
+	if u.Payload != nil {
+		payloadSize = u.Payload.Size()
 	}
 
-	return references, nil
+	// TransactionType
+	return serializer.OneByte +
+		// NetworkID
+		serializer.UInt64ByteSize +
+		// CreationSlot
+		SlotIndexLength +
+		u.ContextInputs.Size() +
+		u.Inputs.Size() +
+		// InputsCommitment
+		InputsCommitmentLength +
+		u.Outputs.Size() +
+		u.Allotments.Size() +
+		payloadSize
 }
 
-func (t *Transaction) BICInputs() ([]*BlockIssuanceCreditInput, error) {
-	references := make([]*BlockIssuanceCreditInput, 0, len(t.Essence.ContextInputs))
-	for _, input := range t.Essence.ContextInputs {
-		switch castInput := input.(type) {
-		case *BlockIssuanceCreditInput:
-			references = append(references, castInput)
-		case *CommitmentInput, *RewardInput:
-			// ignore this type
-		default:
-			return nil, ErrUnknownContextInputType
-		}
+// syntacticallyValidate checks whether the transaction essence is syntactically valid.
+// The function does not syntactically validate the input or outputs themselves.
+func (u *Transaction) syntacticallyValidate(protoParams ProtocolParameters) error {
+	expectedNetworkID := protoParams.NetworkID()
+	if u.NetworkID != expectedNetworkID {
+		return ierrors.Wrapf(ErrTxEssenceNetworkIDInvalid, "got %v, want %v (%s)", u.NetworkID, expectedNetworkID, protoParams.NetworkName())
 	}
 
-	return references, nil
-}
-
-func (t *Transaction) RewardInputs() ([]*RewardInput, error) {
-	references := make([]*RewardInput, 0, len(t.Essence.ContextInputs))
-	for _, input := range t.Essence.ContextInputs {
-		switch castInput := input.(type) {
-		case *RewardInput:
-			references = append(references, castInput)
-		case *CommitmentInput, *BlockIssuanceCreditInput:
-			// ignore this type
-		default:
-			return nil, ErrUnknownContextInputType
-		}
-	}
-
-	return references, nil
-}
-
-// Returns the first commitment input in the transaction if it exists or nil.
-func (t *Transaction) CommitmentInput() *CommitmentInput {
-	for _, input := range t.Essence.ContextInputs {
-		switch castInput := input.(type) {
-		case *BlockIssuanceCreditInput, *RewardInput:
-			// ignore this type
-		case *CommitmentInput:
-			return castInput
-		default:
-			return nil
-		}
-	}
-
-	return nil
-}
-
-func (t *Transaction) Size() int {
-	// PayloadType
-	return serializer.TypeDenotationByteSize +
-		t.Essence.Size() +
-		t.Unlocks.Size()
-}
-
-func (t *Transaction) String() string {
-	// TODO: stringify for debugging purposes
-	return fmt.Sprintf("Transaction[%v, %v]", t.Essence, t.Unlocks)
-}
-
-// syntacticallyValidate syntactically validates the Transaction.
-func (t *Transaction) syntacticallyValidate(api API) error {
-	// limit unlock block count = input count
-	if len(t.Unlocks) != len(t.Essence.Inputs) {
-		return ierrors.Errorf("unlock block count must match inputs in essence, %d vs. %d", len(t.Unlocks), len(t.Essence.Inputs))
-	}
-
-	if err := t.Essence.syntacticallyValidate(api.ProtocolParameters()); err != nil {
-		return ierrors.Errorf("transaction essence is invalid: %w", err)
-	}
-
-	if err := ValidateUnlocks(t.Unlocks,
-		UnlocksSigUniqueAndRefValidator(api),
+	if err := SyntacticallyValidateContextInputs(u.ContextInputs,
+		ContextInputsSyntacticalUnique(),
 	); err != nil {
-		return ierrors.Errorf("invalid unlocks: %w", err)
+		return err
 	}
 
-	return nil
+	if err := SyntacticallyValidateInputs(u.Inputs,
+		InputsSyntacticalUnique(),
+		InputsSyntacticalIndicesWithinBounds(),
+	); err != nil {
+		return err
+	}
+
+	return SyntacticallyValidateOutputs(u.Outputs,
+		OutputsSyntacticalDepositAmount(protoParams),
+		OutputsSyntacticalExpirationAndTimelock(),
+		OutputsSyntacticalNativeTokens(),
+		OutputsSyntacticalChainConstrainedOutputUniqueness(),
+		OutputsSyntacticalFoundry(),
+		OutputsSyntacticalAccount(),
+		OutputsSyntacticalNFT(),
+		OutputsSyntacticalDelegation(),
+	)
 }
 
-func (t *Transaction) WorkScore(workScoreStructure *WorkScoreStructure) (WorkScore, error) {
-	// we account for the network traffic only on "Payload" level
-	workScoreEssenceData, err := workScoreStructure.DataByte.Multiply(t.Size())
+func (u *Transaction) WorkScore(workScoreStructure *WorkScoreStructure) (WorkScore, error) {
+	workScoreContextInputs, err := u.ContextInputs.WorkScore(workScoreStructure)
 	if err != nil {
 		return 0, err
 	}
 
-	workScoreEssence, err := t.Essence.WorkScore(workScoreStructure)
+	workScoreInputs, err := u.Inputs.WorkScore(workScoreStructure)
 	if err != nil {
 		return 0, err
 	}
 
-	workScoreUnlocks, err := t.Unlocks.WorkScore(workScoreStructure)
+	workScoreOutputs, err := u.Outputs.WorkScore(workScoreStructure)
 	if err != nil {
 		return 0, err
 	}
 
-	return workScoreEssenceData.Add(workScoreEssence, workScoreUnlocks)
+	workScoreAllotments, err := u.Allotments.WorkScore(workScoreStructure)
+	if err != nil {
+		return 0, err
+	}
+
+	var workScorePayload WorkScore
+	if u.Payload != nil {
+		workScorePayload, err = u.Payload.WorkScore(workScoreStructure)
+		if err != nil {
+			return 0, err
+		}
+	}
+
+	return workScoreContextInputs.Add(workScoreInputs, workScoreOutputs, workScoreAllotments, workScorePayload)
 }
