+++ vm/stardust/vm_stardust_test.go
@@ -76,8 +76,8 @@ func TestNFTTransition(t *testing.T) {
 	nftAddr := iotago.NFTAddressFromOutputID(inputIDs[0])
 	nftID := nftAddr.NFTID()
 
-	essence := &iotago.TransactionEssence{
-		TransactionInputEssence: &iotago.TransactionInputEssence{
+	essence := &iotago.Transaction{
+		TransactionEssence: &iotago.TransactionEssence{
 			Inputs: inputIDs.UTXOInputs(),
 		},
 		Outputs: iotago.TxEssenceOutputs{
@@ -95,8 +95,8 @@ func TestNFTTransition(t *testing.T) {
 	sigs, err := essence.Sign(testAPI, inputIDs.OrderedSet(inputs.OutputSet()).MustCommitment(testAPI), ident1AddrKeys)
 	require.NoError(t, err)
 
-	tx := &iotago.Transaction{
-		Essence: essence,
+	tx := &iotago.SignedTransaction{
+		Transaction: essence,
 		Unlocks: iotago.Unlocks{
 			&iotago.SignatureUnlock{Signature: sigs[0]},
 		},
@@ -163,8 +163,8 @@ func TestCirculatingSupplyMelting(t *testing.T) {
 		},
 	}
 
-	essence := &iotago.TransactionEssence{
-		TransactionInputEssence: &iotago.TransactionInputEssence{
+	essence := &iotago.Transaction{
+		TransactionEssence: &iotago.TransactionEssence{
 			Inputs: inputIDs.UTXOInputs(),
 		},
 		Outputs: iotago.TxEssenceOutputs{
@@ -201,8 +201,8 @@ func TestCirculatingSupplyMelting(t *testing.T) {
 	sigs, err := essence.Sign(testAPI, inputIDs.OrderedSet(inputs.OutputSet()).MustCommitment(testAPI), ident1AddrKeys)
 	require.NoError(t, err)
 
-	tx := &iotago.Transaction{
-		Essence: essence,
+	tx := &iotago.SignedTransaction{
+		Transaction: essence,
 		Unlocks: iotago.Unlocks{
 			&iotago.SignatureUnlock{Signature: sigs[0]},
 			&iotago.ReferenceUnlock{Reference: 0},
@@ -221,7 +221,7 @@ func TestStardustTransactionExecution(t *testing.T) {
 		name           string
 		vmParams       *vm.Params
 		resolvedInputs vm.ResolvedInputs
-		tx             *iotago.Transaction
+		tx             *iotago.SignedTransaction
 		wantErr        error
 	}
 	tests := []test{
@@ -509,8 +509,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 			}
 
 			creationSlot := iotago.SlotIndex(750)
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: creationSlot,
 				},
@@ -712,8 +712,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 					InputSet:        inputs,
 					CommitmentInput: &iotago.Commitment{Index: creationSlot},
 				},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						// basic
 						&iotago.SignatureUnlock{Signature: sigs[0]},
@@ -782,8 +782,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -809,8 +809,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						// should be an AccountUnlock
@@ -851,8 +851,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					CreationSlot: 110,
 					ContextInputs: iotago.TxEssenceContextInputs{
 						&iotago.BlockIssuanceCreditInput{
@@ -897,8 +897,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs, BlockIssuanceCreditInputSet: bicInputs, CommitmentInput: commitmentInput},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -934,8 +934,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					CreationSlot: 110,
 					ContextInputs: iotago.TxEssenceContextInputs{
 						&iotago.BlockIssuanceCreditInput{
@@ -980,8 +980,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs, BlockIssuanceCreditInputSet: bicInputs, CommitmentInput: commitmentInput},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -1016,8 +1016,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					CreationSlot: 110,
 					ContextInputs: iotago.TxEssenceContextInputs{
 						&iotago.BlockIssuanceCreditInput{
@@ -1053,8 +1053,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs, BlockIssuanceCreditInputSet: bicInputs, CommitmentInput: commitmentInput},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -1090,8 +1090,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					CreationSlot: 110,
 					ContextInputs: iotago.TxEssenceContextInputs{
 						&iotago.BlockIssuanceCreditInput{
@@ -1127,8 +1127,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs, BlockIssuanceCreditInputSet: bicInputs, CommitmentInput: commitment},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -1163,8 +1163,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					CreationSlot: 110,
 					Inputs:       inputIDs.UTXOInputs(),
 				},
@@ -1191,8 +1191,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs, CommitmentInput: commitment},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -1227,8 +1227,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					CreationSlot: 110,
 					Inputs:       inputIDs.UTXOInputs(),
 				},
@@ -1260,8 +1260,8 @@ func TestStardustTransactionExecution(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -1342,8 +1342,8 @@ func runStardustTransactionExecutionTest(t *testing.T, test *txExecTest) {
 		}
 
 		// create the transaction essence
-		txEssence := &iotago.TransactionEssence{
-			TransactionInputEssence: &iotago.TransactionInputEssence{
+		txEssence := &iotago.Transaction{
+			TransactionEssence: &iotago.TransactionEssence{
 				NetworkID:     testProtoParams.NetworkID(),
 				CreationSlot:  100,
 				ContextInputs: iotago.TxEssenceContextInputs{},
@@ -1360,16 +1360,16 @@ func runStardustTransactionExecutionTest(t *testing.T, test *txExecTest) {
 		// pass the signatures and get the unlock conditions
 		unlocks := test.unlocksFunc(sigs, testAddresses)
 
-		tx := &iotago.Transaction{
-			Essence: txEssence,
-			Unlocks: unlocks,
+		tx := &iotago.SignedTransaction{
+			Transaction: txEssence,
+			Unlocks:     unlocks,
 		}
 
 		txBytes, err := testAPI.Encode(tx, serix.WithValidation())
 		require.NoError(t, err)
 
 		// we deserialize to be sure that all serix rules are applied (like lexically ordering or multi addresses)
-		tx = &iotago.Transaction{}
+		tx = &iotago.SignedTransaction{}
 		_, err = testAPI.Decode(txBytes, tx, serix.WithValidation())
 		require.NoError(t, err)
 
@@ -2532,7 +2532,7 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 		name           string
 		vmParams       *vm.Params
 		resolvedInputs vm.ResolvedInputs
-		tx             *iotago.Transaction
+		tx             *iotago.SignedTransaction
 		wantErr        error
 	}
 	tests := []test{
@@ -2631,8 +2631,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 			}
 
 			creationSlot := iotago.SlotIndex(10)
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: creationSlot,
 				},
@@ -2663,8 +2663,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 						Index: iotago.SlotIndex(0),
 					},
 				},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						&iotago.ReferenceUnlock{Reference: 0},
@@ -2695,8 +2695,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 			}
@@ -2712,8 +2712,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -2744,8 +2744,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 			}
@@ -2759,8 +2759,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						&iotago.SignatureUnlock{Signature: sigs[0]},
@@ -2795,8 +2795,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 			}
@@ -2810,8 +2810,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						&iotago.ReferenceUnlock{Reference: 0},
@@ -2845,8 +2845,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 			}
@@ -2860,8 +2860,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						&iotago.ReferenceUnlock{Reference: 0},
@@ -2897,8 +2897,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 			}
@@ -2910,8 +2910,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.NFTUnlock{Reference: 1},
 						&iotago.NFTUnlock{Reference: 0},
@@ -2941,8 +2941,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 			}
 
 			creationSlot := iotago.SlotIndex(5)
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: creationSlot,
 				},
@@ -2962,8 +2962,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 						Index: iotago.SlotIndex(0),
 					},
 				},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -2992,8 +2992,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 			}
 
 			creationSlot := iotago.SlotIndex(10)
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: creationSlot,
 				},
@@ -3013,8 +3013,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 						Index: creationSlot,
 					},
 				},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -3068,8 +3068,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 			}
@@ -3083,8 +3083,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						&iotago.AccountUnlock{Reference: 0},
@@ -3123,8 +3123,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -3148,8 +3148,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						&iotago.AccountUnlock{Reference: 0},
@@ -3195,8 +3195,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -3222,8 +3222,8 @@ func TestTxSemanticInputUnlocks(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						// should be an AccountUnlock
@@ -3253,7 +3253,7 @@ func TestTxSemanticDeposit(t *testing.T) {
 		name           string
 		vmParams       *vm.Params
 		resolvedInputs vm.ResolvedInputs
-		tx             *iotago.Transaction
+		tx             *iotago.SignedTransaction
 		wantErr        error
 	}
 	tests := []test{
@@ -3308,8 +3308,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 			}
 
 			creationSlot := iotago.SlotIndex(5)
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: creationSlot,
 				},
@@ -3343,8 +3343,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 						Index: creationSlot,
 					},
 				},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						&iotago.ReferenceUnlock{Reference: 0},
@@ -3374,8 +3374,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -3411,8 +3411,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -3435,8 +3435,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: 5,
 				},
@@ -3458,8 +3458,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -3482,8 +3482,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: 5,
 				},
@@ -3505,8 +3505,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -3540,8 +3540,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 			}
 
 			creationSlot := iotago.SlotIndex(5)
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: creationSlot,
 				},
@@ -3568,8 +3568,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 						Index: creationSlot,
 					},
 				},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -3597,8 +3597,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -3625,8 +3625,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -3654,8 +3654,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -3686,8 +3686,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -3715,8 +3715,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -3746,8 +3746,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -3782,8 +3782,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -3816,8 +3816,8 @@ func TestTxSemanticDeposit(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -3874,7 +3874,7 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 		name           string
 		vmParams       *vm.Params
 		resolvedInputs vm.ResolvedInputs
-		tx             *iotago.Transaction
+		tx             *iotago.SignedTransaction
 		wantErr        error
 	}
 	tests := []test{
@@ -3905,8 +3905,8 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -3926,9 +3926,9 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
-					Unlocks: iotago.Unlocks{},
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
+					Unlocks:     iotago.Unlocks{},
 				},
 				wantErr: nil,
 			}
@@ -3955,8 +3955,8 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 				}
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -3981,9 +3981,9 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
-					Unlocks: iotago.Unlocks{},
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
+					Unlocks:     iotago.Unlocks{},
 				},
 				wantErr: nil,
 			}
@@ -4006,8 +4006,8 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -4027,9 +4027,9 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
-					Unlocks: iotago.Unlocks{},
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
+					Unlocks:     iotago.Unlocks{},
 				},
 				wantErr: iotago.ErrMaxNativeTokensCountExceeded,
 			}
@@ -4051,8 +4051,8 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 				}
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -4071,9 +4071,9 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
-					Unlocks: iotago.Unlocks{},
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
+					Unlocks:     iotago.Unlocks{},
 				},
 				wantErr: iotago.ErrMaxNativeTokensCountExceeded,
 			}
@@ -4106,8 +4106,8 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 				}
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: outs,
@@ -4119,9 +4119,9 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
-					Unlocks: iotago.Unlocks{},
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
+					Unlocks:     iotago.Unlocks{},
 				},
 				wantErr: iotago.ErrMaxNativeTokensCountExceeded,
 			}
@@ -4155,8 +4155,8 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 				}
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: outputs,
@@ -4168,9 +4168,9 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
-					Unlocks: iotago.Unlocks{},
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
+					Unlocks:     iotago.Unlocks{},
 				},
 				wantErr: nil,
 			}
@@ -4216,8 +4216,8 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: outputs,
@@ -4229,9 +4229,9 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
-					Unlocks: iotago.Unlocks{},
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
+					Unlocks:     iotago.Unlocks{},
 				},
 				wantErr: iotago.ErrMaxNativeTokensCountExceeded,
 			}
@@ -4259,8 +4259,8 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 			amountToModify := cpyNativeTokens[ntCount/2].Amount
 			amountToModify.Add(amountToModify, big.NewInt(1))
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -4280,9 +4280,9 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
-					Unlocks: iotago.Unlocks{},
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
+					Unlocks:     iotago.Unlocks{},
 				},
 				wantErr: iotago.ErrNativeTokenSumUnbalanced,
 			}
@@ -4321,8 +4321,8 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 			cpyNativeTokens := nativeTokens.Clone()
 			cpyNativeTokens = append(cpyNativeTokens, tpkg.RandNativeToken())
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -4350,9 +4350,9 @@ func TestTxSemanticNativeTokens(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
-					Unlocks: iotago.Unlocks{},
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
+					Unlocks:     iotago.Unlocks{},
 				},
 				wantErr: iotago.ErrNativeTokenSumUnbalanced,
 			}
@@ -4377,7 +4377,7 @@ func TestTxSemanticOutputsSender(t *testing.T) {
 		name           string
 		vmParams       *vm.Params
 		resolvedInputs vm.ResolvedInputs
-		tx             *iotago.Transaction
+		tx             *iotago.SignedTransaction
 		wantErr        error
 	}
 	tests := []test{
@@ -4406,8 +4406,8 @@ func TestTxSemanticOutputsSender(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -4480,8 +4480,8 @@ func TestTxSemanticOutputsSender(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						&iotago.ReferenceUnlock{Reference: 0},
@@ -4505,8 +4505,8 @@ func TestTxSemanticOutputsSender(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -4530,8 +4530,8 @@ func TestTxSemanticOutputsSender(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -4559,8 +4559,8 @@ func TestTxSemanticOutputsSender(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -4586,8 +4586,8 @@ func TestTxSemanticOutputsSender(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -4617,8 +4617,8 @@ func TestTxSemanticOutputsSender(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -4645,8 +4645,8 @@ func TestTxSemanticOutputsSender(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -4674,8 +4674,8 @@ func TestTxSemanticOutputsSender(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -4701,8 +4701,8 @@ func TestTxSemanticOutputsSender(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -4729,7 +4729,7 @@ func TestTxSemanticOutputsIssuer(t *testing.T) {
 		name           string
 		vmParams       *vm.Params
 		resolvedInputs vm.ResolvedInputs
-		tx             *iotago.Transaction
+		tx             *iotago.SignedTransaction
 		wantErr        error
 	}
 	tests := []test{
@@ -4762,8 +4762,8 @@ func TestTxSemanticOutputsIssuer(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -4795,8 +4795,8 @@ func TestTxSemanticOutputsIssuer(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						&iotago.SignatureUnlock{Signature: sigs[1]},
@@ -4840,8 +4840,8 @@ func TestTxSemanticOutputsIssuer(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -4953,8 +4953,8 @@ func TestTxSemanticOutputsIssuer(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						&iotago.SignatureUnlock{Signature: sigs[1]},
@@ -4992,8 +4992,8 @@ func TestTxSemanticOutputsIssuer(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 				},
 				Outputs: iotago.TxEssenceOutputs{
@@ -5025,8 +5025,8 @@ func TestTxSemanticOutputsIssuer(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						&iotago.SignatureUnlock{Signature: sigs[1]},
@@ -5054,7 +5054,7 @@ func TestTxSemanticTimelocks(t *testing.T) {
 		name           string
 		vmParams       *vm.Params
 		resolvedInputs vm.ResolvedInputs
-		tx             *iotago.Transaction
+		tx             *iotago.SignedTransaction
 		wantErr        error
 	}
 	tests := []test{
@@ -5076,8 +5076,8 @@ func TestTxSemanticTimelocks(t *testing.T) {
 			}
 
 			creationSlot := iotago.SlotIndex(10)
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: creationSlot,
 				},
@@ -5096,8 +5096,8 @@ func TestTxSemanticTimelocks(t *testing.T) {
 						Index: creationSlot,
 					},
 				},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -5124,8 +5124,8 @@ func TestTxSemanticTimelocks(t *testing.T) {
 			}
 
 			creationSlot := iotago.SlotIndex(10)
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: 10,
 				},
@@ -5144,8 +5144,8 @@ func TestTxSemanticTimelocks(t *testing.T) {
 						Index: creationSlot,
 					},
 				},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -5172,8 +5172,8 @@ func TestTxSemanticTimelocks(t *testing.T) {
 			}
 
 			creationSlot := iotago.SlotIndex(666)
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: creationSlot,
 				},
@@ -5192,8 +5192,8 @@ func TestTxSemanticTimelocks(t *testing.T) {
 						Index: creationSlot,
 					},
 				},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -5220,8 +5220,8 @@ func TestTxSemanticTimelocks(t *testing.T) {
 			}
 
 			creationSlot := iotago.SlotIndex(1005)
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: creationSlot,
 				},
@@ -5237,8 +5237,8 @@ func TestTxSemanticTimelocks(t *testing.T) {
 				resolvedInputs: vm.ResolvedInputs{
 					InputSet: inputs,
 				},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -5266,7 +5266,7 @@ func TestTxSemanticMana(t *testing.T) {
 		name           string
 		vmParams       *vm.Params
 		resolvedInputs vm.ResolvedInputs
-		tx             *iotago.Transaction
+		tx             *iotago.SignedTransaction
 		wantErr        error
 	}
 	tests := []test{
@@ -5287,8 +5287,8 @@ func TestTxSemanticMana(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: 10 + 100*testProtoParams.ParamEpochDurationInSlots(),
 				},
@@ -5324,8 +5324,8 @@ func TestTxSemanticMana(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -5350,8 +5350,8 @@ func TestTxSemanticMana(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs: inputIDs.UTXOInputs(),
 					Allotments: iotago.Allotments{
 						&iotago.Allotment{Value: 50},
@@ -5391,8 +5391,8 @@ func TestTxSemanticMana(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -5417,8 +5417,8 @@ func TestTxSemanticMana(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: 15,
 				},
@@ -5441,8 +5441,8 @@ func TestTxSemanticMana(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -5467,8 +5467,8 @@ func TestTxSemanticMana(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: 15,
 				},
@@ -5491,8 +5491,8 @@ func TestTxSemanticMana(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -5527,8 +5527,8 @@ func TestTxSemanticMana(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: 15,
 				},
@@ -5551,8 +5551,8 @@ func TestTxSemanticMana(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 						&iotago.ReferenceUnlock{Reference: 0},
@@ -5578,8 +5578,8 @@ func TestTxSemanticMana(t *testing.T) {
 				},
 			}
 
-			essence := &iotago.TransactionEssence{
-				TransactionInputEssence: &iotago.TransactionInputEssence{
+			essence := &iotago.Transaction{
+				TransactionEssence: &iotago.TransactionEssence{
 					Inputs:       inputIDs.UTXOInputs(),
 					CreationSlot: 15,
 				},
@@ -5609,8 +5609,8 @@ func TestTxSemanticMana(t *testing.T) {
 					API: testAPI,
 				},
 				resolvedInputs: vm.ResolvedInputs{InputSet: inputs},
-				tx: &iotago.Transaction{
-					Essence: essence,
+				tx: &iotago.SignedTransaction{
+					Transaction: essence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sigs[0]},
 					},
@@ -5667,8 +5667,8 @@ func TestManaRewardsClaimingStaking(t *testing.T) {
 		},
 	}
 
-	essence := &iotago.TransactionEssence{
-		TransactionInputEssence: &iotago.TransactionInputEssence{
+	essence := &iotago.Transaction{
+		TransactionEssence: &iotago.TransactionEssence{
 			Inputs: inputIDs.UTXOInputs(),
 		},
 		Outputs: iotago.TxEssenceOutputs{
@@ -5700,8 +5700,8 @@ func TestManaRewardsClaimingStaking(t *testing.T) {
 	sigs, err := essence.Sign(testAPI, inputIDs.OrderedSet(inputs.OutputSet()).MustCommitment(testAPI), identAddrKeys)
 	require.NoError(t, err)
 
-	tx := &iotago.Transaction{
-		Essence: essence,
+	tx := &iotago.SignedTransaction{
+		Transaction: essence,
 		Unlocks: iotago.Unlocks{
 			&iotago.SignatureUnlock{Signature: sigs[0]},
 		},
@@ -5746,8 +5746,8 @@ func TestManaRewardsClaimingDelegation(t *testing.T) {
 	}
 	delegationID := iotago.DelegationIDFromOutputID(inputIDs[0])
 
-	essence := &iotago.TransactionEssence{
-		TransactionInputEssence: &iotago.TransactionInputEssence{
+	essence := &iotago.Transaction{
+		TransactionEssence: &iotago.TransactionEssence{
 			Inputs:       inputIDs.UTXOInputs(),
 			CreationSlot: currentSlot,
 		},
@@ -5766,8 +5766,8 @@ func TestManaRewardsClaimingDelegation(t *testing.T) {
 	sigs, err := essence.Sign(testAPI, inputIDs.OrderedSet(inputs.OutputSet()).MustCommitment(testAPI), identAddrKeys)
 	require.NoError(t, err)
 
-	tx := &iotago.Transaction{
-		Essence: essence,
+	tx := &iotago.SignedTransaction{
+		Transaction: essence,
 		Unlocks: iotago.Unlocks{
 			&iotago.SignatureUnlock{Signature: sigs[0]},
 		},
@@ -6071,7 +6071,7 @@ func TestTxSemanticAddressRestrictions(t *testing.T) {
 		},
 	}
 
-	makeTransaction := func(output iotago.Output) (vm.InputSet, iotago.Signature, *iotago.TransactionEssence) {
+	makeTransaction := func(output iotago.Output) (vm.InputSet, iotago.Signature, *iotago.Transaction) {
 		inputIDs := tpkg.RandOutputIDs(1)
 
 		inputs := vm.InputSet{
@@ -6086,8 +6086,8 @@ func TestTxSemanticAddressRestrictions(t *testing.T) {
 			},
 		}
 
-		essence := &iotago.TransactionEssence{
-			TransactionInputEssence: &iotago.TransactionInputEssence{
+		essence := &iotago.Transaction{
+			TransactionEssence: &iotago.TransactionEssence{
 				Inputs:       inputIDs.UTXOInputs(),
 				CreationSlot: 10,
 			},
@@ -6114,8 +6114,8 @@ func TestTxSemanticAddressRestrictions(t *testing.T) {
 				}
 
 				resolvedInputs := vm.ResolvedInputs{InputSet: inputs}
-				tx := &iotago.Transaction{
-					Essence: transactionEssence,
+				tx := &iotago.SignedTransaction{
+					Transaction: transactionEssence,
 					Unlocks: iotago.Unlocks{
 						&iotago.SignatureUnlock{Signature: sig},
 					},
